# Skip parsing not used json fields