… go.sum
… pkg/scraper/types_easyjson.go
… vendor/k8s.io/kubernetes/pkg/kubelet/apis/stats/v1alpha1/go.mod
… vendor/k8s.io/kubernetes/pkg/kubelet/apis/stats/v1alpha1/types_easyjson.go