/cc @s-urbaniak 

Based on work benchmarking whole scraping decoding json is most costly operation. Skipping parsing those fields reduces scraping usage of CPU by 17% and allocs by 78%. 

For now I added a micro benchmark just for unmarshaling. It shows reduction of  CPU usage by 47% and allocs by 87%.
```
name             old time/op    new time/op    delta
JSONUnmarshal-8    55.0µs ± 1%    28.7µs ± 7%  -47.83%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
JSONUnmarshal-8    12.2kB ± 0%     5.3kB ± 0%  -56.17%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
JSONUnmarshal-8       184 ± 0%        23 ± 0%  -87.50%  (p=0.000 n=10+10)
```